Various nodewatcher module changes.
Wireless reporting module now uses iw for all reporting as
it is the most reliable. VPN reporting module has been
removed as it is obsolete now that we have tunneldigger.
The traffic module has been renamed to interfaces and
upgraded with MTU reporting (QoS reporting following in a
later upgrade).